+++ CHANGELOG.md
@@ -1,4 +1,3 @@
 # ChangeLog
 
-* fix multiblock in-world preview crash in some cases.
-* fix mekanism chemical trait capability of multiple slots have a incorrect behaviour.
+* fix mekanism chemical filter doesn't work.